create tests for Resource, too
